+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed [here](https://github.com/sensu-plugins/community/blob/master/HOW_WE_CHANGELOG.md)
 
 ## [Unreleased]
+### Added
+- metrics-alb.rb: get metrics for the Application Load Balancer (v2 API) (@multani)
 
 ## [11.5.0] - 2018-05-17
 ### Added